Add a private constructor for FinnishLocaleUtil.
FinnishLocaleUtil is not an ordinary class, but only
a collection of static helper methods. It is not meant
to be instantiated. But if we don't want holes in our
code coverage because the constructor is not invoked, we
have to create some class which does with reflection all
things that normal code never does and is not supposed to do.
Here I named it InvokeImpossibleThingsForCodeCoverageTest.